First cut at trying to make Android version with equivalent JS API
